+++ ladderly-io/src/server/api/root.ts
@@ -3,7 +3,7 @@ import { userRouter } from './routers/user'
 import { authRouter } from './routers/auth'
 import { checklistRouter } from './routers/checklist'
 import { chatRouter } from './routers/chat'
-import { jobSearchRouter } from './routers/jobSearch'
+import { jobSearchRouter } from './routers/jobSearch/router'
 import { courseRouter } from './routers/course'
 import { quizRouter } from './routers/quiz'
 import { certificateRouter } from './routers/certificate'
